Separate state components from stateless components
