+++ src/components/NotFound.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const NotFound = () => (
+    <div className="home-container">
+        <h1>Not Found</h1>
+    </div>
+);
+
+export default NotFound;
